refactored apis

+++ server/ctrls/course-ctrls.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import { Course } from "../models.js";
 
 export const get_course = async (req, res) => {
@@ -10,14 +11,11 @@ export const get_course = async (req, res) => {
       throw new Error("unauthorized");
     } else {
       // get courses
-      const { query } = req.body;
-      if (query) {
-        const courses = await Course.find(query);
-        res.status(200).send({ courses, message: "courses found" });
-      } else {
-        res.status(404);
-        throw new Error("courses not found");
-      }
+      const query = req.query;
+      // check if _id is present and convert it to ObjectId
+      if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+      const courses = await Course.find(query);
+      res.status(200).send({ data: courses, message: "courses found" });
     }
   } catch (err) {
     if (res.statusCode < 400) res.status(500);
@@ -62,13 +60,20 @@ export const edit_course = async (req, res) => {
     } else {
       // update course
       const { edits, query } = req.body;
-      const result = await Course.updateMany(query, { $set: edits }, { new: true });
-      // check if course updated
-      if (!result) {
+      if (query) {
+        // check if _id is present and convert it to ObjectId
+        if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+        const result = await Course.updateMany(query, { $set: edits }, { new: true });
+        // check if course updated
+        if (!result) {
+          res.status(404);
+          throw new Error("courses not found");
+        } else {
+          res.status(201).send({ courses: result, message: "courses updated" });
+        }
+      } else {
         res.status(404);
         throw new Error("courses not found");
-      } else {
-        res.status(201).send({ courses: result, message: "courses updated" });
       }
     }
   } catch (err) {
@@ -88,13 +93,20 @@ export const delete_course = async (req, res) => {
     } else {
       // delete courses
       const { query } = req.body;
-      const result = await Course.deleteMany(query);
-      // check if courses deleted
-      if (!result) {
+      if (query) {
+        // check if _id is present and convert it to ObjectId
+        if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+        const result = await Course.deleteMany(query);
+        // check if courses deleted
+        if (!result) {
+          res.status(404);
+          throw new Error("courses not found");
+        } else {
+          res.status(202).send({ courses: result, message: "courses deleted" });
+        }
+      } else {
         res.status(404);
         throw new Error("courses not found");
-      } else {
-        res.status(202).send({ message: "courses deleted" });
       }
     }
   } catch (err) {
