+++ server/ctrls/test-ctrls.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import { Test } from "../models.js";
 
 export const get_test = async (req, res) => {
@@ -10,14 +11,11 @@ export const get_test = async (req, res) => {
       throw new Error("unauthorized");
     } else {
       // get tests
-      const { query } = req.body;
-      if (query) {
-        const tests = await Test.find(query);
-        res.status(200).send({ tests, message: "tests found" });
-      } else {
-        res.status(404);
-        throw new Error("tests not found");
-      }
+      const query = req.query;
+      // check if _id is present and convert it to ObjectId
+      if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+      const tests = await Test.find(query);
+      res.status(200).send({ data: tests, message: "tests found" });
     }
   } catch (err) {
     if (res.statusCode < 400) res.status(500);
@@ -61,13 +59,20 @@ export const edit_test = async (req, res) => {
     } else {
       // update test
       const { edits, query } = req.body;
-      const result = await Test.updateMany(query, { $set: edits }, { new: true });
-      // check if test updated
-      if (!result) {
+      if (query) {
+        // check if _id is present and convert it to ObjectId
+        if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+        const result = await Test.updateMany(query, { $set: edits }, { new: true });
+        // check if test updated
+        if (!result) {
+          res.status(404);
+          throw new Error("tests not found");
+        } else {
+          res.status(201).send({ tests: result, message: "tests updated" });
+        }
+      } else {
         res.status(404);
         throw new Error("tests not found");
-      } else {
-        res.status(201).send({ tests: result, message: "tests updated" });
       }
     }
   } catch (err) {
@@ -87,13 +92,20 @@ export const delete_test = async (req, res) => {
     } else {
       // delete tests
       const { query } = req.body;
-      const result = await Test.deleteMany(query);
-      // check if tests deleted
-      if (!result) {
+      if (query) {
+        // check if _id is present and convert it to ObjectId
+        if (query._id) query._id = new mongoose.Types.ObjectId(query._id);
+        const result = await Test.deleteMany(query);
+        // check if tests deleted
+        if (!result) {
+          res.status(404);
+          throw new Error("tests not found");
+        } else {
+          res.status(202).send({ tests: result, message: "tests deleted" });
+        }
+      } else {
         res.status(404);
         throw new Error("tests not found");
-      } else {
-        res.status(202).send({ message: "tests deleted" });
       }
     }
   } catch (err) {
